more cleanup

+++ transcribe-wrapped.js
@@ -2,15 +2,33 @@ const which = require("which");
 const spawn = require('child_process').spawn;
 // const spawn = require('await-spawn')
 const fs = require('fs-extra');
-const moment = require("moment-timezone");
-const momentDurationFormatSetup = require("moment-duration-format");
-
 const whisperPath = which.sync('whisper')
 
 l = console.log;
 
-const thing = moment.duration(123, "months").format();
+// /usr/bin/python3 /usr/local/bin/whisper uploads/0
+/**
+ * Translates seconds into human readable format of seconds, minutes, hours, days, and years
+ *
+ * @param  {number} seconds The number of seconds to be processed
+ * @return {string}         The phrase describing the amount of time
+ */
+function forHumans ( seconds ) {
+  var levels = [
+    [Math.floor(seconds / 31536000), 'years'],
+    [Math.floor((seconds % 31536000) / 86400), 'days'],
+    [Math.floor(((seconds % 31536000) % 86400) / 3600), 'hours'],
+    [Math.floor((((seconds % 31536000) % 86400) % 3600) / 60), 'minutes'],
+    [(((seconds % 31536000) % 86400) % 3600) % 60, 'seconds'],
+  ];
+  var returntext = '';
 
+  for (var i = 0, max = levels.length; i < max; i++) {
+    if ( levels[i][0] === 0 ) continue;
+    returntext += ' ' + levels[i][0] + ' ' + (levels[i][0] === 1 ? levels[i][1].substr(0, levels[i][1].length-1): levels[i][1]);
+  };
+  return returntext.trim();
+}
 
 async function transcribe(filename, path, language, model, websocketNumber){
   return new Promise((resolve, reject) => {
@@ -42,17 +60,12 @@ async function transcribe(filename, path, language, model, websocketNumber){
       l('transcribe arguments');
       l(arguments);
 
-      // save date when starting just to see how long it's taking
-      const startingDate = new Date();
-      l(startingDate);
-
       // load websocket by passed number
       let websocketConnection;
       if(global.ws[websocketNumber]){
         websocketConnection = global.ws[websocketNumber]
       }
 
-      l(global.ws);
 
       const ls = spawn(whisperPath, arguments);
 
@@ -69,33 +82,57 @@ async function transcribe(filename, path, language, model, websocketNumber){
       });
 
 
+      // save date when starting just to see how long it's taking
+      const startingDate = new Date();
+      l(startingDate);
+
       ls.on('close', code => {
-        // delete original upload to save space
-        const shouldDeleteOriginalUpload = false;
-        if(shouldDeleteOriginalUpload){
-          const originalUpload = `./uploads/${splitFilename}`;
-          fs.unlinkSync(originalUpload);
+        l('code');
+        l(code);
+
+        if(code === 0){
+          // delete original upload to save space
+          const shouldDeleteOriginalUpload = false;
+          if(shouldDeleteOriginalUpload){
+            const originalUpload = `./uploads/${splitFilename}`;
+            fs.unlinkSync(originalUpload);
+          }
+
+          // where the transcription was saved
+          const containingDir = `./transcriptions/${splitFilename}`;
+
+          // transcribed srt file
+          const transcribedSrtFile = `${containingDir}/${filename}.srt`;
+
+          // copy srt with the original filename
+          (async function(){
+            await fs.copy(`${containingDir}/${splitFilename}.srt`, transcribedSrtFile)
+          })()
+
+          // tell frontend upload is done
+          websocketConnection.send(JSON.stringify({
+            status: 'Completed',
+            url: transcribedSrtFile
+          }), function () {});
+
+          // return await
+          resolve(code);
+
+          const secondsDifference = (new Date() - startingDate) / 1000
+          const humanReadableTime = forHumans(Math.round(secondsDifference));
+
+          const outputText = `
+            filename: ${filename}
+            secondsDifference: ${secondsDifference}
+            humanReadableTime: ${humanReadableTime}
+            language: ${language}
+            model: ${model}
+          `.replace(/^ +/gm, ''); // remove indentation
+
+          fs.appendFileSync(`${containingDir}/processing_data.txt`, outputText, 'utf8');
+        } else {
+          l('FAILED!');
         }
-
-        // where the transcription was saved
-        const containingDir = `./transcriptions/${splitFilename}`;
-
-        // transcribed srt file
-        const transcribedSrtFile = `${containingDir}/${filename}.srt`;
-
-        // copy srt with the original filename
-        (async function(){
-          await fs.copy(`${containingDir}/${splitFilename}.srt`, transcribedSrtFile)
-        })()
-
-        // tell frontend upload is done
-        websocketConnection.send(JSON.stringify({
-          status: 'Completed',
-          url: transcribedSrtFile
-        }), function () {});
-
-        // return await
-        resolve(code);
 
         console.log(`child process exited with code ${code}`);
       });
